+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -514,5 +514,24 @@ namespace Dfe.Complete.Api.Controllers
             await sender.Send(request, cancellationToken);
             return NoContent();
         }
+        
+        /// <summary>
+        /// Confirm the incoming trust has completed all actions for transfer project.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/IncomingTrustHasCompleteAllActions")]
+        [SwaggerResponse(204, "Updated the confirm incoming trust has completed all actions task successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project not found.")]
+        public async Task<IActionResult> UpdateConfirmIncomingTrustHasCompleteAllActionsTaskAsync(
+            [FromBody] UpdateConfirmIncomingTrustHasCompleteAllActionsTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
     }
 }